requirements.gemini.txt
flask
python-dotenv
Werkzeug
Pillow
google-genai
fastapi
uvicorn[standard]